Disable installing/upgrading incompatible extension versions
